server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jarvis-server"
version = "0.1.0"
description = "Jarvis AI Chief of Staff - Server Component"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jarvis Team"}
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "qdrant-client>=1.9.0",
    "structlog>=24.0.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.0.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "easyocr>=1.7",
    "pytesseract>=0.3.10",
    "fastembed>=0.4",
    "arq>=0.26",
    "redis>=5.0",
    "orjson>=3.9",
    "python-dateutil>=2.8",
    "pillow>=10.0",
    # Google Calendar API
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth>=2.30.0",
    # LLM API
    "anthropic>=0.39.0",
    # Audio transcription
    "faster-whisper>=1.0.0",
    # Web UI
    "jinja2>=3.1.0",
    # HTML parsing for email body extraction
    "beautifulsoup4>=4.12.0",
    "markdown>=3.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
]

[project.scripts]
jarvis-server = "jarvis_server.main:run"
jarvis-worker = "arq:main"

[tool.hatch.build.targets.wheel]
packages = ["src/jarvis_server"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
